pyproject.toml
[tool.poetry]
name = "psycopg2-pgevents"
version = "0.2.0"
description = "PostGreSQL LISTEN/NOTIFY functionality, via psycopg2"
readme = "README.rst"
homepage = "https://github.com/shawalli/psycopg2-pgevents"
repository = "https://github.com/shawalli/psycopg2-pgevents"
authors = ["Shawn Wallis <shawn.p.wallis@gmail.com>"]
license = "MIT"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
include = [
  "LICENSE.txt",
]

[tool.poetry.dependencies]
python = "^3.7"
psycopg2-binary = ">=2.7.5"
pytest = "^5.4.3"
pytest-cov = "^2.10.0"
coveralls = "^2.1.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.6.0"
testfixtures = "^6.14.1"

[tool.black]
line-length = 120
target-version = ["py37", "py38"]
include = "\\.pyi?$"
exclude = """
/(
    \\.eggs
  | \\.git
  | \\.hg
  | \\.mypy_cache
  | \\.tox
  | \\.venv
  | _build
  | buck-out
  | build
  | dist
)/
"""

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
